Support nested Parameters.
Support nested Parameters by adding a type assertion to the query
loop in evaluationStage.go. Remove the check for unexported
accessors in the parser, and replace with a check for unexported
field access in evaluationStage.

This allows nesting maps in an arbitrary fashion with the
MapParameters type, as shown by the unit test.

Fixes #70